+++ src/ngrest/plugins/Datetime.php
@@ -3,6 +3,7 @@
 namespace luya\admin\ngrest\plugins;
 
 use luya\admin\ngrest\base\Plugin;
+use Yii;
 
 /**
  * Date and Time input field
@@ -13,7 +14,7 @@ use luya\admin\ngrest\base\Plugin;
  * Example empty Date configuration
  *
  * ```
- * ['timestamp', ['Datetime', 'emptyMessage' => 'No Date']],
+ * ['timestamp', ['Datetime', 'emptyMessage' => 'No Date', 'format' => 'dd.MM.yyyy']],
  * ```
  *
  * @author Basil Suter <basil@nadar.io>
@@ -22,22 +23,30 @@ use luya\admin\ngrest\base\Plugin;
 class Datetime extends Plugin
 {
     /**
-     * @var string This text will be displayed in the list overview when no date has been slected
+     * @var string This text will be displayed in the list overview when no date has been selected
      * or date is null/empty.
      */
     public $emptyMessage = '-';
+    
+    /**
+     * @var string Use custom datetime format by [date filter](https://docs.angularjs.org/api/ng/filter/date). Default is 'short'. Use false to take \yii\i18n\Formatter::$datetimeFormat as fallback.
+     * @since 2.0.0
+     */
+    public $format = 'short';
     
     /**
      * @inheritdoc
      */
     public function renderList($id, $ngModel)
     {
+        $format = $this->format ?? Yii::$app->formatter->datetimeFormat;
+        
         return [
-            $this->createTag('span', null, ['ng-show' => $ngModel, 'ng-bind' => $ngModel.'*1000 | date : \'short\'']),
+            $this->createTag('span', null, ['ng-show' => $ngModel, 'ng-bind' => $ngModel."*1000 | date : '$format'"]),
             $this->createTag('span', $this->emptyMessage, ['ng-show' => '!'.$ngModel]),
         ];
     }
-
+    
     /**
      * @inheritdoc
      */
@@ -45,7 +54,7 @@ class Datetime extends Plugin
     {
         return $this->createFormTag('zaa-datetime', $id, $ngModel);
     }
-
+    
     /**
      * @inheritdoc
      */